+++ bean-searcher-boot-starter/src/main/java/com/ejlchina/searcher/boot/BeanSearcherAutoConfiguration.java
@@ -157,7 +157,7 @@ public class BeanSearcherAutoConfiguration {
 	@Bean
 	@ConditionalOnMissingBean(SizeLimitParamFilter.class)
 	public SizeLimitParamFilter sizeLimitParamFilter(BeanSearcherProperties config) {
-		return new SizeLimitParamFilter(config.getParams().getFilter().getMaxAllowedSize());
+		return new SizeLimitParamFilter(config.getParams().getFilter().getMaxParaMapSize());
 	}
 
 	@Bean